Move helper type to its own file
